Refactor transpile main validation
